refactor(ultros/web): extract WebState and static-file handlers into submodules
web.rs has grown to 1322 lines mixing router assembly, ~50 handlers,
state plumbing, and static-file serving. First wave of splitting:

- web/state.rs (new) — WebState struct + 12 FromRef impls. Pure
  state plumbing; no logic. The router stays in web.rs and consumes
  the type via `pub(crate) use self::state::WebState`.

- web/static_files.rs (new) — favicon/robots/static_path,
  get_item_icon, fallback_item_icon, plus the private get_file
  helper and the include_dir-backed/disk-backed get_static_file
  variants for release/debug builds. Self-contained: only depends
  on web::error.

Net: web.rs is down ~200 lines and its imports thin out
correspondingly. No behavior change. The remaining handlers
(retainers, lists, characters, groups, invites) are still inline
in web.rs — a follow-up can pull those out per resource if the
churn is worth it.

Includes a cargo fmt pass on ultros-db/src/lists.rs that was
already failing fmt on main.
